requirements.txt
cdsapi
cftime
matplotlib
netCDF4
numpy
openpyxl
optuna
pandas
pyarrow
python-box
pytorch
pytorch_geometric
scikit-learn
scipy
tqdm
wandb
xarray
xgboost